refactor(#55): flatten books api structure
